fix(router): copy over data during data resolution
